requirements/main.txt
Django==3.0.3
celery==4.4.0
djangorestframework== 3.11.0
psycopg2==2.8.4
redis==3.4.1
